버그 수정

+++ StereoMix-Launcher/src/FileHelper.cs
@@ -266,7 +266,10 @@ public static class FileHelper
                 await HttpHelper.SaveToFile(window, response, fileStream);
             });
             
-            Directory.Delete(window.InstallDirectory, true);
+            if (Directory.Exists(window.InstallDirectory))
+            {
+                Directory.Delete(window.InstallDirectory, true);
+            }
             
             ZipFile.ExtractToDirectory(tempFile, window.InstallDirectory, true);
             File.Delete(tempFile);
@@ -298,7 +301,10 @@ public static class FileHelper
                 await HttpHelper.DevSaveToFile(window, response, fileStream);
             });
             
-            Directory.Delete(window.DevInstallDirectory, true);
+            if (Directory.Exists(window.DevInstallDirectory))
+            {
+                Directory.Delete(window.DevInstallDirectory, true);
+            }
             
             ZipFile.ExtractToDirectory(tempFile, window.DevInstallDirectory, true);
             File.Delete(tempFile);
